refactor: rename some types
