+++ src/client/services/MediaDatabaseService.ts
@@ -1,5 +1,5 @@
 import { $ } from "@qwik.dev/core";
-import type { MediaFileRecord } from "../../types/FileRecord";
+import type { MediaFileRecord } from "../../types/Records";
 import { IndexedDatabaseService } from "~/client/services";
 
 export const MediaDatabaseService = $((store: string) =>
