--- src/types/IndexedDatabaseRecord.ts
@@ -1,6 +1,0 @@
-export interface IndexedDatabaseRecord {
-  /**
-   * The unique identifier of the file record.
-   */
-  id: string;
-}
