+++ src/types/Records.ts
@@ -1,5 +1,11 @@
-import type { IndexedDatabaseRecord } from "./IndexedDatabaseRecord";
 import type { ImageMetadata, MediaMetadata, VideoMetadata } from "./MediaMetadata";
+
+export interface IndexedDatabaseRecord {
+  /**
+   * The unique identifier of the file record.
+   */
+  id: string;
+}
 
 /**
  * Represents an IndexedDB record of a file.
@@ -19,14 +25,14 @@ export interface FileRecord extends IndexedDatabaseRecord {
  * Represents an IndexedDB record of an image file.
  */
 export interface ImageFileRecord extends FileRecord {
-  meta: ImageMetadata
+  meta: ImageMetadata;
 }
 
 /**
  * Represents an IndexedDB record of a video file.
  */
 export interface VideoFileRecord extends FileRecord {
-  meta: VideoMetadata
+  meta: VideoMetadata;
 }
 
 /** Represents an IndexedDB media file record, which can be either an image or a video. */
